pyproject.toml
[project]
name = "fastlet"
version = "0.1.0"
description = ""
authors = [{ name = "draew6" }]
dependencies = [
    "uvicorn[standard]==0.34.2",
    "python-jose==3.5.0",
    "fastapi==0.115.12",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "python-multipart==0.0.20",
    "bcrypt==4.3.0",
    "sendgrid==6.12.3",
    "psycopg-binary==3.2.9",
    "git+https://github.com/draew6/piping_bag.git@v0.2",
]
requires-python = ">=3.13"

[project.scripts]
fastroutes = "fastlet.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
